[Fix] TestReview의 푼 문제수 타입 오류 해결

+++ src/main/java/kr/sols/domain/testReview/dto/TestReivewDataLabDto.java
@@ -16,7 +16,7 @@ public class TestReivewDataLabDto {
     private String trPosition;
     private String trCareer;
     private int trProblemNum;
-    private int trSolvedNum;
+    private float trSolvedNum;
     private String trPassStatus;
 
     public static TestReivewDataLabDto fromEntity(TestReview testReview) {
