+++ src/main/java/kr/sols/domain/testReview/dto/TestReviewDto.java
@@ -25,7 +25,7 @@ public class TestReviewDto {
     private String trPosition;
     private String trCareer;
     private int trProblemNum;
-    private int trSolvedNum;
+    private float trSolvedNum;
     private String trPassStatus;
     private List<String> trProblemType;
     private String trComment;
